+++ src/components/BulkEditList/BulkEditListResult/BulkEditInApp/forms/utils.js
@@ -12,3 +12,9 @@ export const handleAdd = ({ getFilteredFields, fields, fieldTemplate, getDefault
 
   setFields(finalizedFields);
 };
+
+export const isAddButtonShown = (index, fields, options) => {
+  return index === fields.length - 1 && fields.length !== options.length - 1;
+};
+
+export const isActionsVisible = (field) => field.option;